refactor: 장학 제도 SSR로 리팩터링
