+++ app/[locale]/academics/helper/courses/CourseCard.tsx
