+++ app/[locale]/academics/helper/courses/CourseDetailModal.tsx
