+++ app/[locale]/academics/helper/courses/CourseList.tsx
