+++ app/[locale]/academics/helper/courses/CourseListHeader.tsx
