+++ app/[locale]/academics/helper/courses/CourseListRow.tsx
