+++ app/[locale]/academics/helper/courses/CourseRow.tsx
